feat: add comprehensive tests for Plot2D and PlotMesh functionality
